Projet final/Bac2 + TFB/Mercury-TFB--main/requirements.txt
numpy
pandas
scikit-learn
scipy
plotly
matplotlib
tqdm
Pillow